requirements.txt
streamlit>=1.41.0
pandas>=2.0
plotly>=5.0
google-genai>=1.0.0
